created function for arrow listener
